requirements.txt
langchain-groq 
langchain-core 
streamlit 
langchain-chroma 
langchain-nomic 
langchain
nomic 
python-dotenv 
langchain-community 
rank_bm25 
cohere 
nomic[local] 
streamlit-feedback 
qdrant-client 
langchain-cohere
langchain-huggingface

